web_scraper_and_flask_app/web_scraper/requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.0.0
h11==0.14.0
idna==3.4
numpy==1.23.4
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.1
psycopg2-binary==2.9.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyVirtualDisplay==3.0
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sqlalchemy==1.4.42
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
